+++ src/main/webapp/GestionDesTaches.jsp
@@ -166,7 +166,7 @@
                           <a href="EditTache?idTache=<%= temp.getIdTache() %>"class="btn edit">
                                 <ion-icon name="create-outline"></ion-icon> 
                             </a>
-                            <a type="button" href="DeleteTache?idTache=<%= temp.getIdTache() %>&idProjet=<%= temp.getIdProjet() %>" class="btn delete">
+                            <a type="button" href="DeleteTache?idTache=<%= temp.getIdTache() %>&idProjet=<%= temp.getIdProjet() %>" class="btn delete" onclick="return confirmDelete();">
                                 <ion-icon name="trash-outline"></ion-icon> 
                             </a>
